refactor(#349): chunk 4 — collapse SEC/coverage service files
Chunk 4 of the 2026-04-19 research-tool refocus §1.1. Mechanical
consolidation of retired service files into their two target modules.

Merged into app/services/fundamentals.py (1870 lines, 4 sections):
- Section 1: fundamentals snapshot (was fundamentals.py)
- Section 2: XBRL fact storage + ingestion runs (was financial_facts.py)
- Section 3: Period normalization (was financial_normalization.py)
- Section 4: SEC change-driven planner/executor (was sec_incremental.py)

Merged into app/services/coverage.py (1869 lines, 3 sections):
- Section 1: Tier management (was coverage.py)
- Section 2: filings_status classifier (was coverage_audit.py)
- Section 3: Filings backfill (was filings_backfill.py)

Deleted:
- app/services/financial_facts.py
- app/services/financial_normalization.py
- app/services/sec_incremental.py
- app/services/coverage_audit.py
- app/services/filings_backfill.py

Import sites rewritten in 13 files (app/ + tests/) via sed. All top-level
symbols from the retired modules remain importable under the new paths.
Codex verified: no out-of-repo caller, no circular-import risk from
the module-joining, no functional drop. 2143 tests passing, ruff clean.

Net: 8 SEC/coverage service files → 3 (fundamentals.py, coverage.py,
filings.py unchanged). Spec §1.1 target "~8 SEC-related files → 2" hit
(filings.py stays because it owns the generic filings refresh for both
SEC and Companies House).
